+++ test/data/test_stats.py
@@ -7,6 +7,7 @@ from torch_frame.data.stats import StatType, compute_col_stats
 from torch_frame.datasets.fake import _random_timestamp
 from torch_frame.stype import (
     categorical,
+    embedding,
     multicategorical,
     numerical,
     sequence_numerical,
@@ -14,7 +15,7 @@ from torch_frame.stype import (
 )
 
 
-def test_compute_col_stats_all_numerical():
+def test_compute_col_stats_numerical():
     ser = pd.Series([1, 2, 3])
     stype = numerical
     assert compute_col_stats(ser, stype) == {
@@ -24,7 +25,7 @@ def test_compute_col_stats_all_numerical():
     }
 
 
-def test_compute_col_stats_all_categorical():
+def test_compute_col_stats_categorical():
     ser = pd.Series(['a', 'a', 'a', 'b', 'c'])
     stype = categorical
     assert compute_col_stats(ser, stype) == {
@@ -32,7 +33,7 @@ def test_compute_col_stats_all_categorical():
     }
 
 
-def test_compute_col_stats_all_multi_categorical():
+def test_compute_col_stats_multi_categorical():
     for ser in [
             pd.Series(['a|a|b', 'a|c', 'c|a', 'a|b|c', '', None]),
             # # Testing with leading and traling whitespace
@@ -47,7 +48,7 @@ def test_compute_col_stats_all_multi_categorical():
         }
 
 
-def test_compute_col_stats_all_timestamp():
+def test_compute_col_stats_timestamp():
     start_date = datetime(2000, 1, 1)
     end_date = datetime(2023, 1, 1)
     num_rows = 10
@@ -82,7 +83,7 @@ def test_compute_col_stats_all_timestamp():
     assert (year_range[0] == year_range[1] == 1900)
 
 
-def test_compute_col_stats_all_timestamp_with_all_nan():
+def test_compute_col_stats_timestamp_with_nan():
     ser = pd.Series([np.nan, np.nan, np.nan, np.nan, np.nan, np.nan, np.nan])
     stype = timestamp
     year_range = compute_col_stats(ser, stype)[StatType.YEAR_RANGE]
@@ -90,7 +91,7 @@ def test_compute_col_stats_all_timestamp_with_all_nan():
     assert np.isnan(year_range[0]) and np.isnan(year_range[1])
 
 
-def test_compute_col_stats_all_sequence_numerical():
+def test_compute_col_stats_sequence_numerical():
     ser = pd.Series([[1, 2, 3], [4, 5, 6]])
     stype = sequence_numerical
     assert compute_col_stats(ser, stype) == {
@@ -100,7 +101,7 @@ def test_compute_col_stats_all_sequence_numerical():
     }
 
 
-def test_compute_col_stats_all_sequence_numerical_with_nan_inf():
+def test_compute_col_stats_sequence_numerical_with_nan_inf():
     ser = pd.Series([[1, 2, 3, np.nan], [4, 5, 6]])
     expected_col_stats = {
         StatType.MEAN: 3.5,
@@ -113,7 +114,7 @@ def test_compute_col_stats_all_sequence_numerical_with_nan_inf():
     assert compute_col_stats(ser, stype) == expected_col_stats
 
 
-def test_compute_col_stats_all_sequence_numerical_with_all_nan():
+def test_compute_col_stats_sequence_numerical_with_nan():
     ser = pd.Series([[np.nan, np.nan, np.nan, np.inf],
                      [np.nan, np.nan, np.nan]])
     stype = sequence_numerical
@@ -144,11 +145,20 @@ def test_compute_col_stats_numerical_with_nan():
     }
 
 
-def test_compute_col_stats_numerical_all_nan():
+def test_compute_col_stats_numerical_nan():
     ser = pd.Series([np.nan, np.nan, np.nan, np.nan])
     stype = numerical
     assert compute_col_stats(ser, stype) == {
         StatType.MEAN: np.nan,
         StatType.STD: np.nan,
         StatType.QUANTILES: [np.nan, np.nan, np.nan, np.nan, np.nan],
+    }
+
+
+def test_compute_col_stats_embedding():
+    emb_dim = 12
+    ser = pd.Series([np.random.rand(emb_dim) for _ in range(3)])
+    stype = embedding
+    assert compute_col_stats(ser, stype) == {
+        StatType.EMB_DIM: emb_dim,
     }