+++ test/nn/encoder/test_stype_encoder.py
@@ -43,13 +43,14 @@ def test_categorical_feature_encoder(encoder_cls_kwargs):
         **encoder_cls_kwargs[1],
     )
     feat_cat = tensor_frame.feat_dict[stype.categorical]
-    x = encoder(feat_cat)
+    col_names = tensor_frame.col_names_dict[stype.categorical]
+    x = encoder(feat_cat, col_names)
     assert x.shape == (feat_cat.size(0), feat_cat.size(1), 8)
 
     # Perturb the first column
     num_categories = len(stats_list[0][StatType.COUNT])
     feat_cat[:, 0] = (feat_cat[:, 0] + 1) % num_categories
-    x_perturbed = encoder(feat_cat)
+    x_perturbed = encoder(feat_cat, col_names)
     # Make sure other column embeddings are unchanged
     assert (x_perturbed[:, 1:, :] == x[:, 1:, :]).all()
 
@@ -91,7 +92,8 @@ def test_numerical_feature_encoder(encoder_cls_kwargs):
                                     stype=stype.numerical,
                                     **encoder_cls_kwargs[1])
     feat_num = tensor_frame.feat_dict[stype.numerical]
-    x = encoder(feat_num)
+    col_names = tensor_frame.col_names_dict[stype.numerical]
+    x = encoder(feat_num, col_names)
     assert x.shape == (feat_num.size(0), feat_num.size(1), 8)
     if "post_module" in encoder_cls_kwargs[1]:
         assert encoder.post_module is not None
@@ -100,7 +102,7 @@ def test_numerical_feature_encoder(encoder_cls_kwargs):
 
     # Perturb the first column
     feat_num[:, 0] = feat_num[:, 0] + 10.0
-    x_perturbed = encoder(feat_num)
+    x_perturbed = encoder(feat_num, col_names)
     # Make sure other column embeddings are unchanged
     assert (x_perturbed[:, 1:, :] == x[:, 1:, :]).all()
 
@@ -136,14 +138,15 @@ def test_multicategorical_feature_encoder(encoder_cls_kwargs):
         **encoder_cls_kwargs[1],
     )
     feat_multicat = tensor_frame.feat_dict[stype.multicategorical]
-    x = encoder(feat_multicat)
+    col_names = tensor_frame.col_names_dict[stype.multicategorical]
+    x = encoder(feat_multicat, col_names)
     assert x.shape == (feat_multicat.size(0), feat_multicat.size(1), 8)
 
     # Perturb the first column
     num_categories = len(stats_list[0][StatType.MULTI_COUNT])
     feat_multicat[:,
                   0].values = (feat_multicat[:, 0].values + 1) % num_categories
-    x_perturbed = encoder(feat_multicat)
+    x_perturbed = encoder(feat_multicat, col_names)
     # Make sure other column embeddings are unchanged
     assert (x_perturbed[:, 1:, :] == x[:, 1:, :]).all()
 
@@ -169,8 +172,9 @@ def test_categorical_feature_encoder_with_nan(encoder_cls_kwargs):
         **encoder_cls_kwargs[1],
     )
     feat_cat = tensor_frame.feat_dict[stype.categorical]
+    col_names = tensor_frame.col_names_dict[stype.categorical]
     isnan_mask = feat_cat == -1
-    x = encoder(feat_cat)
+    x = encoder(feat_cat, col_names)
     assert x.shape == (feat_cat.size(0), feat_cat.size(1), 8)
     # Make sure there's no NaNs in x
     assert (~torch.isnan(x)).all()
@@ -198,8 +202,9 @@ def test_numerical_feature_encoder_with_nan(encoder_cls_kwargs):
                                     stype=stype.numerical,
                                     **encoder_cls_kwargs[1])
     feat_num = tensor_frame.feat_dict[stype.numerical]
+    col_names = tensor_frame.col_names_dict[stype.numerical]
     isnan_mask = feat_num.isnan()
-    x = encoder(feat_num)
+    x = encoder(feat_num, col_names)
     assert x.shape == (feat_num.size(0), feat_num.size(1), 8)
     # Make sure there's no NaNs in x
     assert (~torch.isnan(x)).all()
@@ -236,8 +241,9 @@ def test_multicategorical_feature_encoder_with_nan(encoder_cls_kwargs):
         **encoder_cls_kwargs[1],
     )
     feat_multicat = tensor_frame.feat_dict[stype.multicategorical]
+    col_names = tensor_frame.col_names_dict[stype.multicategorical]
     isnan_mask = feat_multicat.values == -1
-    x = encoder(feat_multicat)
+    x = encoder(feat_multicat, col_names)
     assert x.shape == (feat_multicat.size(0), feat_multicat.size(1), 8)
     # Make sure there's no NaNs in x
     assert (~torch.isnan(x)).all()
@@ -252,11 +258,9 @@ def test_text_embedded_encoder():
     dataset = FakeDataset(
         num_rows=num_rows,
         stypes=[
-            torch_frame.numerical,
-            torch_frame.categorical,
             torch_frame.text_embedded,
         ],
-        text_embedder_cfg=TextEmbedderConfig(
+        col_to_text_embedder_cfg=TextEmbedderConfig(
             text_embedder=HashTextEmbedder(text_emb_channels),
             batch_size=None),
     )
@@ -271,11 +275,42 @@ def test_text_embedded_encoder():
         stats_list=stats_list,
         stype=stype.text_embedded,
     )
-    x_text = tensor_frame.feat_dict[stype.text_embedded]
-    x = encoder(x_text)
-    assert x.shape == (
+    feat_text = tensor_frame.feat_dict[stype.text_embedded]
+    col_names = tensor_frame.col_names_dict[stype.text_embedded]
+    feat = encoder(feat_text, col_names)
+    assert feat.shape == (
         num_rows,
         len(tensor_frame.col_names_dict[stype.text_embedded]),
+        out_channels,
+    )
+
+
+def test_embedding_encoder():
+    num_rows = 20
+    out_channels = 5
+    dataset = FakeDataset(
+        num_rows=num_rows,
+        stypes=[
+            torch_frame.embedding,
+        ],
+    )
+    dataset.materialize()
+    tensor_frame = dataset.tensor_frame
+    stats_list = [
+        dataset.col_stats[col_name]
+        for col_name in tensor_frame.col_names_dict[stype.embedding]
+    ]
+    encoder = LinearEmbeddingEncoder(
+        out_channels=out_channels,
+        stats_list=stats_list,
+        stype=stype.embedding,
+    )
+    feat_text = tensor_frame.feat_dict[stype.embedding]
+    col_names = tensor_frame.col_names_dict[stype.embedding]
+    x = encoder(feat_text, col_names)
+    assert x.shape == (
+        num_rows,
+        len(tensor_frame.col_names_dict[stype.embedding]),
         out_channels,
     )
 
@@ -288,8 +323,6 @@ def test_text_tokenized_encoder():
     dataset = FakeDataset(
         num_rows=num_rows,
         stypes=[
-            torch_frame.numerical,
-            torch_frame.categorical,
             torch_frame.text_tokenized,
         ],
         text_tokenizer_cfg=TextTokenizerConfig(
@@ -315,8 +348,9 @@ def test_text_tokenized_encoder():
         in_channels=text_emb_channels,
         model=model,
     )
-    x_text = tensor_frame.feat_dict[stype.text_tokenized]
-    x = encoder(x_text)
+    feat_text = tensor_frame.feat_dict[stype.text_tokenized]
+    col_names = tensor_frame.col_names_dict[stype.text_tokenized]
+    x = encoder(feat_text, col_names)
     assert x.shape == (
         num_rows,
         len(tensor_frame.col_names_dict[stype.text_tokenized]),