+++ test/nn/models/test_ft_transformer.py
@@ -1,16 +1,18 @@
+import pytest
+
 from torch_frame.data.dataset import Dataset
 from torch_frame.datasets import FakeDataset
 from torch_frame.nn import FTTransformer
 
 
-def test_ft_transformer():
-    batch_size = 10
+@pytest.mark.parametrize('batch_size', [0, 5])
+def test_ft_transformer(batch_size):
     channels = 8
     out_channels = 1
     num_layers = 3
     dataset: Dataset = FakeDataset(num_rows=10, with_nan=False)
     dataset.materialize()
-    tensor_frame = dataset.tensor_frame
+    tensor_frame = dataset.tensor_frame[:batch_size]
     # Feature-based embeddings
     model = FTTransformer(
         channels=channels,