+++ test/utils/test_infer_stype.py
@@ -9,7 +9,7 @@ from torch_frame.utils import infer_df_stype
 
 def get_fake_dataset(
     num_rows: int,
-    text_embedder_cfg: TextEmbedderConfig,
+    col_to_text_embedder_cfg: TextEmbedderConfig,
     with_nan: bool,
 ) -> FakeDataset:
     stypes = [
@@ -23,7 +23,7 @@ def get_fake_dataset(
     dataset = FakeDataset(
         num_rows=num_rows,
         stypes=stypes,
-        text_embedder_cfg=text_embedder_cfg,
+        col_to_text_embedder_cfg=col_to_text_embedder_cfg,
         with_nan=with_nan,
     )
     return dataset
@@ -32,7 +32,8 @@ def get_fake_dataset(
 @pytest.mark.parametrize("with_nan", [True, False])
 def test_infer_df_stype(with_nan):
     num_rows = 200
-    text_embedder_cfg = TextEmbedderConfig(text_embedder=HashTextEmbedder(8))
-    dataset = get_fake_dataset(num_rows, text_embedder_cfg, with_nan)
+    col_to_text_embedder_cfg = TextEmbedderConfig(
+        text_embedder=HashTextEmbedder(8))
+    dataset = get_fake_dataset(num_rows, col_to_text_embedder_cfg, with_nan)
     col_to_stype_inferred = infer_df_stype(dataset.df)
     assert col_to_stype_inferred == dataset.col_to_stype